make loader, transformer as interfaces
